delete_collection
