+++ response/delete_collection.go
